feat: App.tsx code was moved to AppShell.tsx in one refactor. Moved it back and deleted AppShell.tsx.
